+++ example/lib/drop_down_demo2.dart
@@ -76,7 +76,11 @@ class _DropDownDemoState extends State<DropDownDemo2>
       body: Column(children: [
         DropDownHeader(
           controller: dropDownController,
+          boxStyle: DropDownBoxStyle(
+            expand: false,
+          ),
           itemStyle: const DropDownItemStyle(
+            width: 80,
             activeIconColor: Colors.blue,
             activeTextStyle: TextStyle(color: Colors.blue),
           ),
@@ -88,6 +92,12 @@ class _DropDownDemoState extends State<DropDownDemo2>
               activeIcon: const Icon(Icons.arrow_drop_up),
             ),
           ),
+          itemBuilder: (context, index) {
+            if (index == 3) {
+              return Container(width: 200, color: Colors.red);
+            }
+            return null;
+          },
         ),
         Expanded(
           child: Stack(